tools/trap: Refactor the trap tool to make the script easy to understand
1. We refactor the trap script into multiple files. Earlier, there were only two files: ramdumpParser.py and debugsymbolviewer.py
2. We divide the script content into multiple files: trap.py, dumpParser.py, logParser.py, heapNode.py and utils.py
   - trap.py        : It is the wrapper script which contains usage API, initializing code and make instance of relevant class that provides APIs to parse the log file and ram dump
   - logParser.py   : It has logParser class which provide APIs (i.e. find crash point, call stack details, heap corruption, etc) to parse the assert/crash log file
   - dumpParser.py  : It has dumpParser class which has APIs mainly for parsing RAM/Flash dump
   - heapNode.py    : It contains dumpNode class which helps in parsing heap corruption log.
   - utils.py       : It act as utility for all above files and provide API for setting up symbol table and retrieving symbol based on address and vice versa.
3. We split the trap folder into two parts:
   - cli : contains trap tool for command line.
   - gui : contains trap tool for GUI.
4. We remove the debugsymbolviewer.py and add it's content to above script files.
5. Usage for TRAP tool: python3 trap.py -t <log-file> -c <config-path> -b <bin-path> -e <elf-file-name>
   Instead of ramdumpParser.py, we use trap.py as script.
