+++ tools/trap/gui/trap_utils.c
